Fixed bug while parsing manga pages. Fixed bug while pages merge with different width.

+++ downloaders/strategies/MergeFramesStrategy.py
@@ -5,6 +5,9 @@ import os
 
 class MergeFramesStrategy(DownloadStrategy):
 
+    def __init__(self, save_original):
+        self.__save_original = save_original
+
     def Execute(self, path: str):
 
         files = [os.path.join(path, f) for f in os.listdir(path) if f.endswith((".png", ".jpg", ".jpeg"))]
@@ -15,18 +18,22 @@ class MergeFramesStrategy(DownloadStrategy):
 
         images = [Image.open(f) for f in files]
 
-        width = images[0].width
+        global_width = max(images, key=lambda i: i.width).width
         total_height = sum(img.height for img in images)
 
-        result = Image.new("RGB", (width, total_height))
+        result = Image.new("RGBA", (global_width, total_height))
         y_offset = 0
         for img in images:
-            result.paste(img, (0, y_offset))
-            y_offset += img.height
-        result.save(os.path.join(path, "merged.jpg"))
 
-        for f in files:
-            try:
-                os.remove(f)
-            except Exception as e:
-                print(f"Не вдалося видалити {f}: {e}")
+            current_x_pos = (global_width - img.width) // 2
+
+            result.paste(img, (current_x_pos, y_offset))
+            y_offset += img.height
+        result.save(os.path.join(path, "merged.png"))
+
+        if not self.__save_original:
+            for f in files:
+                try:
+                    os.remove(f)
+                except Exception as e:
+                    print(f"Не вдалося видалити {f}: {e}")
